# openapi.yml
openapi: 3.0.0
info:
  title: Library API
  version: 1.0.0
  description: Essa é a documentação da API de livraria do projeto de POO - UNIDOMBOSCO ADS 2024

servers:
  - url: http://localhost:5000
    description: Dev

paths:
  /v1/auth/login:
    post:
      summary: Realiza o login do usuario na plataforma
      tags:
        - "Auth"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hash:
                  type: string
                  example: a035552ecc2f6d88e797faa3497a3cb3d4148299ca5f56bd5a3d8ee29d765f4b
      responses:
        '200':
          description: Usuário logado com sucesso
        '401':
          description: Falha na autenticação

  /v1/user/register:
    post:
      summary: Cria o registro do usuario na base
      tags:
        - "Register"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: example@example.com.br 
                first_name:
                  type: string
                  example: Lorem
                last_name:
                  type: string
                  example: Ipsum
                cellphone:
                  type: string
                  example: 4199999999                  
                password:
                  type: string
                  example: password 
                birth_date:
                  type: string
                  example: 2021-01-01
                address:
                  type: string
                  example: Rua dos mercadores
                number:
                  type: string
                  example: 100                   
                zip_code:
                  type: string
                  example: 00000-000 
                role:
                  type: string
                  example: user                                     
      responses:
        '200':
          description: Usuário criado com sucesso
        '401':
          description: Falha na criação

  /v1/books/query:
    get:
      summary: Lista todos os livros
      tags:
        - "Books"      
      description: Retorna uma lista de todos os livros disponíveis.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'

  /v1/books/{book_id}:
    get:
      summary: Retorna um livro específico
      description: Retorna os detalhes de um livro específico com base no ID fornecido.
      tags:
        - "Books"  
      parameters:
        - name: book_id
          in: path
          description: ID do livro a ser recuperado
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'

  /v1/books/{book_id}/delete:
    delete:
      summary: Exclui um livro
      description: Exclui um livro específico com base no ID fornecido.
      tags:
        - "Books"  
      parameters:
        - name: book_id
          in: path
          description: ID do livro a ser recuperado
          required: true
      responses:
        200:
          description: Livro excluído com sucesso!
        404:
          description: Livro não encontrado

  /v1/books/add:
    post:
      summary: Adiciona um novo livro
      tags:
        - "Books"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Livro adicionado com sucesso
        '400':
          description: Requisição inválida       

  /v1/books/{book_id}/update:
    put:
      summary: Atualiza um livro
      description: Atualiza um livro específico com base no ID fornecido.
      tags:
        - "Books"  
      parameters:
        - name: book_id
          in: path
          description: ID do livro a ser atualizado
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        200:
          description: Livro atualizado com sucesso
        404:
          description: Livro não encontrado
components:
  schemas:
    Book:
      type: object
      properties:
        title:
          type: string
          description: Título do livro
        author:
          type: string
          description: Autor do livro
        publication_date:
          type: string
          format: date
          description: Data de publicação do livro
        isbn:
          type: string
          description: ISBN do livro
        publisher:
          type: string
          description: Editora do livro
        genre:
          type: string
          description: Gênero do livro
        page_count:
          type: integer
          description: Número de páginas do livro
        language:
          type: string
          description: Idioma do livro
        summary:
          type: string
          description: Resumo do livro
        cover_image:
          type: string
          description: URL da imagem de capa do livro
        stock:
          type: integer
          description: Quantidade de exemplares disponíveis em estoque
        price:
          type: number
          format: float
          description: Preço do livro